+++ packages/styled-react/src/index.tsx
@@ -1,4 +1,11 @@
-import {Box, Label as PrimerLabel, type LabelProps as PrimerLabelProps, type SxProp} from '@primer/react'
+import {
+  Box,
+  Label as PrimerLabel,
+  type LabelProps as PrimerLabelProps,
+  type SxProp,
+  ToggleSwitch as PrimerToggleSwitch,
+  type ToggleSwitchProps as PrimerToggleSwitchProps,
+} from '@primer/react'
 import {forwardRef} from 'react'
 import type {
   BackgroundProps,
@@ -32,7 +39,13 @@ const Label = forwardRef(function Label(props, ref) {
   return <Box as={PrimerLabel} ref={ref} {...props} />
 }) as ForwardRefComponent<'span', PrimerLabelProps & Omit<StyledProps, keyof PrimerLabelProps>>
 
-export {Label}
+const ToggleSwitch = forwardRef(function ToggleSwitch(props, ref) {
+  // @ts-expect-error there is an issue with polymorphic `as` with this
+  // component
+  return <Box as={PrimerToggleSwitch} ref={ref} {...props} />
+}) as ForwardRefComponent<'div', PrimerToggleSwitchProps & Omit<StyledProps, keyof PrimerToggleSwitchProps>>
+
+export {Label, ToggleSwitch}
 export {
   ActionList,
   ActionMenu,