+++ packages/styled-react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @primer/styled-react
 
+## 1.0.0-rc.1
+
+### Minor Changes
+
+- [#6627](https://github.com/primer/react/pull/6627) [`6111046`](https://github.com/primer/react/commit/61110467f759c6680797c56d72deb9f86bba4dcd) Thanks [@liuliu-dev](https://github.com/liuliu-dev)! - Update ToggleSwitch component to no longer support sx, add sx wrapper to @primer/styled-react.
+
+- [#6612](https://github.com/primer/react/pull/6612) [`7ad9b81`](https://github.com/primer/react/commit/7ad9b81ac39252438761376cfe8d064294062ff0) Thanks [@joshblack](https://github.com/joshblack)! - Add all components from @primer/react as exports to package
+
+### Patch Changes
+
+- [#6649](https://github.com/primer/react/pull/6649) [`4e1c740`](https://github.com/primer/react/commit/4e1c7409eda128daf9e4f226553699dd0651c843) Thanks [@joshblack](https://github.com/joshblack)! - Update the types output for the `@primer/styled-react` package so that the paths resolve correctly
+
+- Updated dependencies [[`6111046`](https://github.com/primer/react/commit/61110467f759c6680797c56d72deb9f86bba4dcd), [`133d5a5`](https://github.com/primer/react/commit/133d5a5e74b85811d2bb8b536836d5c16680efe1), [`247b3f7`](https://github.com/primer/react/commit/247b3f75c49e16b883e8f0528a036fc62d274ee6), [`de5a4b7`](https://github.com/primer/react/commit/de5a4b7297a44d0dd1ad175ea8d740f570bc27e6), [`3a778b9`](https://github.com/primer/react/commit/3a778b97b93a0fe2c54f585d668cd70d30c0ca56), [`65fc87d`](https://github.com/primer/react/commit/65fc87dc35652c5b228fc7e22d7644645ede2c89), [`3ab94c5`](https://github.com/primer/react/commit/3ab94c5e8853855c8533c8403fd0d0203ab087d6)]:
+  - @primer/react@38.0.0-rc.1
+
+## 1.0.0-rc.0
+
+### Patch Changes
+
+- Updated dependencies [[`bdac258`](https://github.com/primer/react/commit/bdac258a6f610da974ef4b8c25ccef876946fc79), [`3b3cf52`](https://github.com/primer/react/commit/3b3cf52f267da4f44123032bf388dc5ff9f61cf8), [`f4ded58`](https://github.com/primer/react/commit/f4ded585c4f6188390cdc3243018fe63af310633), [`e6c7614`](https://github.com/primer/react/commit/e6c7614d7aaa420ce8518ad54af62e6409fea9dd), [`0c21301`](https://github.com/primer/react/commit/0c21301ba0c7b1d0272258f8fe59026beab83c95), [`1f531cb`](https://github.com/primer/react/commit/1f531cb5c0fb87fc20ab8ce4321367d3f24ab734), [`d6d25dc`](https://github.com/primer/react/commit/d6d25dc9263119103138156158f74b408d300dd2), [`f781f7f`](https://github.com/primer/react/commit/f781f7f5434be4c482a8f7819c73c258b93604ce), [`77a60e7`](https://github.com/primer/react/commit/77a60e7775987ee05b07cd8235ff4a26230b12e2), [`77c8739`](https://github.com/primer/react/commit/77c873936b195915c3f364d01a5b1bb15b0ac1a0), [`661eae0`](https://github.com/primer/react/commit/661eae0a28ee99228400e6c99a483af0523beeb8)]:
+  - @primer/react@38.0.0-rc.0
+
 ## 0.1.0
 
 ### Minor Changes